added signin functionality. Issues with 'proxy' for axios requests where, although the proxy was saved in package.json on the front end, axios was not calling the backend address which was the value of the proxy; it was calling the front end address.  In the end I had to abandon the proxy method and use the full API end point url in the axios request.
Prior to this, I re-created the react front end, using 'vite' because with react cra the proxy was preventing react from even starting (apparently a well known issue).

Created a component to display all math books on axios call, and another to display a single book with details when a specific book image is clicked.

Implemented signup and login functionality, and added react toast for alerting user of the outcome of attempts to signup or login.